typescript, refactoring
