+++ src/styled/GlobalStyles.tsx
@@ -1,0 +1,10 @@
+import { createGlobalStyle } from 'styled-components';
+
+export const Global = createGlobalStyle`
+  * {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    font-family: Alef;
+  }
+`;
