+++ src/styled/StyledButton.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { colors } from './colors';
+import { media } from './media';
 
 const StyledButton = styled.button`
   color: ${colors.backgroundMain};
@@ -12,7 +13,7 @@ const StyledButton = styled.button`
   font-size: 18px;
   margin-top: 20px;
 
-  @media (max-width: 576px) {
+  @media (max-width: ${media.xs}) {
     width: 100%;
     align-self: center;
   }
