+++ src/styled/StyledMainContent.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
 import { colors } from './colors';
+import { media } from './media';
 
 export const StyledMainContent = styled.div`
   border-radius: 5px;
@@ -10,11 +11,11 @@ export const StyledMainContent = styled.div`
   margin: 0 auto;
   display: flex;
   box-shadow: 1px 1px 1px ${colors.secondaryLight};
-  @media (max-width: 768px) {
+  @media (max-width: ${media.sm}) {
     width: 90vw;
     height: 800px;
   }
-  @media (max-width: 576px) {
+  @media (max-width: ${media.xs}) {
     flex-direction: column;
 
     min-width: 100%;
