+++ src/styled/media.tsx
@@ -1,0 +1,5 @@
+export const media = {
+  xs: '576px',
+  sm: '768px',
+  md: '786px',
+};
