+++ src/styled/orderSummaryBlock/StyledOrderSummaryBlock.tsx
@@ -1,12 +1,13 @@
 import styled from 'styled-components';
 import { colors } from '../colors';
+import { media } from '../media';
 
 export const StyledOrderSummaryBlock = styled.div`
   background-color: ${colors.backgroundSecondary};
   width: 40%;
   padding: 30px 10px;
   position: relative;
-  @media (max-width: 576px) {
+  @media (max-width: ${media.xs}) {
     width: 90vw;
     height: 100%;
     text-align: center;
@@ -20,15 +21,12 @@ export const OrderInfoBlock = styled.div`
   height: 70px;
   border-bottom: ${colors.secondary};
   color: ${colors.secondary};
-  @media (max-width: 610px) {
-    padding: 0 10% 0 10%;
-  }
 `;
 export const TotalPrice = styled.div`
   color: ${colors.textMain};
   font-weight: bold;
   padding-top: 10px;
-  @media (max-width: 576px) {
+  @media (max-width: ${media.xs}) {
     padding: 10px 10% 0;
     margin-bottom: 20px;
     order: 3;
