+++ src/styled/orderSummaryBlock/StyledProductInBasket.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
 import { colors } from '../colors';
+import { media } from '../media';
 
 export const StyledProduct = styled.div`
   width: 100%;
@@ -8,19 +9,13 @@ export const StyledProduct = styled.div`
   justify-content: space-between;
   border-bottom: 1px solid grey;
   margin-bottom: 5px;
-  @media (max-width: 700px) {
-    height: 194px;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-  }
-  @media (max-width: 576px) {
+  @media (max-width: ${media.xs}) {
     width: 100%;
     padding: 0 20px;
     height: 160px;
     display: flex;
     flex-direction: row;
+    align-items: center;
     justify-content: space-between;
     border-bottom: 1px solid ${colors.secondary};
     margin-bottom: 5px;
@@ -31,17 +26,13 @@ export const StyledProduct = styled.div`
     padding: 2px;
     width: 50px;
     height: 50px;
-    @media (max-width: 700px) {
+    @media (max-width: ${media.md}) {
       align-self: center;
       border: none;
     }
-    @media (max-width: 576px) {
-      width: 100px;
-      height: 100px;
-    }
-    @media (max-width: 410px) {
-      width: 70px;
-      height: 70px;
+    @media (max-width: ${media.xs}) {
+      width: 85px;
+      height: 85px;
     }
   }
 `;
