+++ tsconfig.json
@@ -6,7 +6,6 @@
             "dom.iterable",
             "esnext"
         ],
-        //test
         "allowJs": true,
         "skipLibCheck": true,
         "esModuleInterop": true,
